requirements.txt
langchain
openai
langchain-openai
PyPDF2
python-dotenv
faiss-cpu
fastapi
uvicorn